% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\docType{data}
\name{schedule}
\alias{schedule}
\title{Schedule Data}
\format{
An object of class \code{data.frame} with 665 rows and 15 columns.
}
\source{
\link{Insert any sources if applicable}
}
\usage{
schedule
}
\description{
This dataset represents a sample class schedule used by several institutions, with some variations in variable names.
Users are encouraged to ensure that the key variables—\code{HRS}, \code{ENRLD}, and \code{INSTRUCTOR}—are formatted correctly for the package to function properly.
}
\details{
A dataset containing the schedule information.

The dataset has \link{number} rows and \link{number} columns, including:
\describe{
\item{HRS}{Credit Hours of a course: int  3 3 3 3 3 3 3 0 3 0 ...}
\item{ENRLD}{Number of students enrolled in a class: int  34 32 34 7 9 8 31 31 12 12 ...}
\item{INSTRUCTOR}{Name of Instructor Teaching the course—real names changed; Char "Bui, Analie" "Calhoun, Alexandra" "Bui, Analie" "Davis, Kheylen" ...}
}
}
\keyword{datasets}
