% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_all_instructors.R
\name{summarize_all_instructors}
\alias{summarize_all_instructors}
\title{Summarize Overload Compensation for All Instructors}
\usage{
summarize_all_instructors(
  schedule_df = schedule,
  L = 4,
  U = 9,
  rate_per_cr = 2500/3,
  reg_load = 12
)
}
\arguments{
\item{schedule_df}{A data frame containing course schedule data with an \code{INSTRUCTOR} column.
Defaults to \code{schedule} if not specified.}

\item{L}{Lower bound of enrollment illegible for peroration (inclusive)}

\item{U}{Upper bound of enrollment illegible for peroration (inclusive)}

\item{rate_per_cr}{Numeric value indicating the rate paid per credit hour.}

\item{reg_load}{A numeric value indicating the regular teaching load (in credit hours).}
}
\value{
A tibble summarizing overload compensation for each instructor,
with formatted compensation and a summary row per instructor.
}
\description{
Applies the overload compensation calculation to each unique instructor
in the schedule and combines the results into one formatted table.
}
\examples{
# summarize_all_instructors(schedule)

}
